Add dynamic type casting support to GetInputValue
This adds an intelligent version of the original to call internal
converters. This also allows converting more complex types like
Transforms to Quaternions (as long as there is a converter for that).

Only works in 2019.2+ since the TypeCache is being used to speed up
any type lookups (native C# assembly diving would be too much of a
performance hit).